Add XML documentation comments
The PixelFormat enum is also split into a separate file and refactored
to follow casing conventions.
