+++ src/Bonsai.Emergent/SerialNumberConverter.cs
@@ -1,9 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Emergent;
 
 namespace Bonsai.Emergent
